pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openmedallion"
version = "0.1.0"
description = "Declarative medallion pipelines in pure open-source Python — local first, cloud portable, fast by default."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "dlt[filesystem,sql_database]>=1.4",
    "sf-hamilton>=1.82",
    "polars>=1.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "websockets>=12.0",
]

[project.optional-dependencies]
s3  = ["s3fs>=2024.1", "boto3>=1.34"]
viz = ["graphviz>=0.20"]
dev = ["pytest>=8"]

[project.scripts]
medallion = "openmedallion.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["openmedallion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
